+++ src/pas/plugins/imio/login.py
@@ -8,9 +8,11 @@ class ImioLoginFormView(BrowserView):
         """Redirect login to authentic"""
         response = self.request.response
         next_url = self.request.get("HTTP_REFERER", self.context.absolute_url())
-        next_url = next_url.replace(api.portal.get().absolute_url(), "")
-        response.redirect(
-            "{0}/authentic-handler/?next_url={1}".format(
+        if next_url:
+            next_url = next_url.replace(api.portal.get().absolute_url(), "")
+            url = "{0}/authentic-handler/?next_url={1}".format(
                 api.portal.get().absolute_url(), next_url
             )
-        )
+        else:
+            url = "{0}/authentic-handler".format(api.portal.get().absolute_url())
+        response.redirect(url)